# [FEATURE]  Optimize `HemeraTermFx` terminal printing (major improvement)