--- a/.gitignore
+++ b/.gitignore
@@ -164,5 +164,5 @@ cython_debug/
 # NyxEngine
 .internal/
 test.py
-benchmark.sh
-profile_data.prof
+.logs
+.logs/