--- a/nyx/hemera_term_fx/hemera_term_fx.py
+++ b/nyx/hemera_term_fx/hemera_term_fx.py
@@ -1,47 +1,6 @@
-"""
-HemeraTermFx Terminal Printing Module:
-
-This module handles the low-level implementation of the terminal printer given an input frame.
-
-The printing workflow is:
-
-    1. Accept a pre-rendered, complete frame (a 2D np.ndarray, dtype=np.uint8)
-
-    2. Convert to a subpixel representation by slicing even/odd rows and stacking them to form a 3D
-       np.ndarray. (Note: a subpixel is a stacked unicode block character "▀", that can be formatted
-       with both a foreground and background color to fit two pixels into one printing character.)
-
-    3. Generate a delta framebuffer of only changed pixels by comparing the new subpixel frame to
-       the previously-cached last frame (or a blank frame if unavailable or otherwise not usable).
-
-    4. Process the delta framebuffer of subpixels into the printing block characters "▀". Colors are
-       issued by ANSI escape sequences. A buffer is held for the last foreground and background
-       colors used to prevent excessive ANSI switching. Cursor relocations are similarly issued only
-       when a non-contiguous subpixel is encountered.
-
-Classes:
-    HemeraTermFx: The primary printing orchestration and processing class. It takes a delta frame-
-        buffer and prints it to the terminal.
-
-Mythology:
-    In the Greek pantheon, Hemera is the personification of day and light, the daughter of Nyx
-    (Night) and Erebus (Darkness). As the embodiment of light she dispels the shadows of night,
-    bringing clarity and illumination to creation and all within it.
-
-Notes:
-
-    - Chars to use: [ █ , ▀ , ▄ ]
-
-    - FPS     Time (ms)
-      =================
-      1         1000.00
-      5          200.00
-      10         100.00
-      15          66.66
-      30          33.33
-"""
-
+import io
 import sys
+from line_profiler import LineProfiler
 import numpy as np
 
 from nyx.hemera_term_fx.term_utils import TerminalUtils
@@ -58,30 +17,50 @@ class HemeraTermFx:
         buffer_log (str): A log of the buffer for debugging purposes.
     """
 
-    def __init__(self, clear_term_on_run: bool = False):
+    def __init__(
+        self,
+        clear_term_on_run: bool = False,
+        # profile_output_file="line_profile_output.txt",
+    ):
         """Construct Hemera with an empty frame buffer."""
         self.old_subpixel_frame: np.ndarray = None
         self.clear_term_on_run: bool = clear_term_on_run
+    #     self.profile_output_file = profile_output_file
+    #     self.profiler = LineProfiler()
+
+    #     # Profile the _generate_string_buffer method
+    #     self.profiler.add_function(self._generate_string_buffer)
+
+    # def _redirect_print_to_profiler(self, *args, **kwargs):
+    #     """Redirect print output to profiler's file output."""
+    #     with open(self.profile_output_file, "a") as f:
+    #         self.profiler.print_stats(stream=f)
 
     def print(self, new_frame: np.ndarray = None):
         """Print the new frame to the terminal in color and using vertical-stacked subpixels.
 
         Args:
-            new_frame (np.ndarray, optional): Completed frame from `AetherRender`. Defaults to None.
+            new_frame (np.ndarray, optional): Completed frame from AetherRender. Defaults to None.
         """
-
         # 1. Generate subpixel frame from the new frame.
         new_subpixel_frame = self._convert_to_subpixels(new_frame)
         # 2. Compare the subpixel frame to the last subpixel frame to get only what has changed.
         delta_frame = self._calculate_delta_framebuffer(new_subpixel_frame)
-        # 3. Print the delta framebuffer to the terminal window.
+
+        # Start profiling the process of printing the frame
+        # self.profiler.enable_by_count()
         self._generate_string_buffer(delta_frame)
+        # self.profiler.disable_by_count()  
+
+        # Log profiler stats into the output file
+        # self._redirect_print_to_profiler()
+
 
     def _convert_to_subpixels(self, new_frame: np.ndarray) -> np.ndarray:
         """Convert the input frame from 2D ndarray -> 3D ndarray of even/odd row pixel colors.
 
         Args:
-            new_frame (np.ndarray, optional): Completed frame from `AetherRender`.
+            new_frame (np.ndarray, optional): Completed frame from AetherRender.
 
         Returns:
             np.ndarray: The subpixel 3D NumPy array.
@@ -126,53 +105,68 @@ class HemeraTermFx:
 
         return delta_frame
 
+    def join_buffer(self, run_buffer):
+        return "".join(run_buffer)
+
+    def write_to_term(self, run_buffer):
+        sys.stdout.write(run_buffer)
+
+    def flush_to_term(self):
+        sys.stdout.flush()
+
+    def sum_bg(self, delta_frame):
+        delta_frame[1] = delta_frame[0] + delta_frame[1]
+        return delta_frame
+    
     def _generate_string_buffer(self, delta_frame: np.ndarray):
         """Convert the delta frame to its color-formatted `str` representation form before printing
         to the terminal.
-
         Args:
             delta_frame (np.ndarray): The delta frame to process and print.
         """
-        run_buffer = ""
+        # Calculate sum of fg + bg
+        delta_frame = self.sum_bg(delta_frame)
+        # Calculate the sum of each row
+        row_sums = np.sum(delta_frame[1], axis=1)
+        # Start the string buffer
+        buffer = io.StringIO()
+        # Initialize loop variables outside the loop
         last_ansi_fg_color, last_ansi_bg_color = np.uint8(0), np.uint8(0)
-        last_subpixel_pair = (np.uint8(0), np.uint8(0))
+        last_subpixel_sum = np.uint8(0)
+        empty_pixel = np.uint8(0)
         _, h, w = delta_frame.shape
-
-        # Iterate the subpixel delta frame using a 2D index
+        fg_color, bg_color = np.uint8(0), np.uint8(0)
+        row_buffer = []
         for y in range(h):
-            row_has_updates = False  # Track if the row has any updates
-            for x in range(w):
-                # Save each axis-0 colors to a tuple
-                fg_color, bg_color = subpixel_pair = tuple(delta_frame[:, y, x])
+            row_buffer = []
+            if row_sums[y] > 0:
+                for x in range(w):
+                    sum_color = delta_frame[1, y, x]
 
-                # Check if subpixel is printable (not (0, 0))
-                if subpixel_pair != (np.uint8(0), np.uint8(0)):
-                    row_has_updates = True  # Mark row as having updates
+                    # Non-zero pixel check (printed pixel)
+                    if sum_color != empty_pixel:
+                        # row_has_updates = True
+                        fg_color = delta_frame[0, y, x]
+                        bg_color = sum_color - fg_color
+                        # Skip cursor movement if it's the same row/column as the last printed pixel
+                        if last_subpixel_sum == empty_pixel:
+                            row_buffer.append(TerminalUtils.cursor_abs_move(x, y))
 
-                    # Issue a cursor relocate and flush the run buffer if the last subpixel pair was
-                    # non-printing (0, 0)
-                    if last_subpixel_pair == (np.uint8(0), np.uint8(0)):
-                        sys.stdout.write(run_buffer)
-                        run_buffer = TerminalUtils.cursor_abs_move(x, y)
+                        # Only write color change sequences when necessary (skip if same as last)
+                        if fg_color != last_ansi_fg_color:
+                            row_buffer.append(f"\033[38;5;{fg_color}m")
+                            last_ansi_fg_color = fg_color
+                        if bg_color != last_ansi_bg_color:
+                            row_buffer.append(f"\033[48;5;{bg_color}m")
+                            last_ansi_bg_color = bg_color
 
-                    # Issue color format sequences only when a new color is needed
-                    if fg_color != last_ansi_fg_color:
-                        run_buffer += f"\033[38;5;{fg_color}m"
-                    if bg_color != last_ansi_bg_color:
-                        run_buffer += f"\033[48;5;{bg_color}m"
+                        # Add the printed character
+                        row_buffer.append("▀")
+                    # Cache the last sum
+                    last_subpixel_sum = sum_color
+                # Add the row buffer for the changed row
+                buffer.write("".join(row_buffer) + "\n")
 
-                    # Add the printing character
-                    run_buffer += "▀"
-
-                # Cache the color values
-                last_subpixel_pair = subpixel_pair
-
-            # Add new line at the end of the row **only if the row had updates**
-            if row_has_updates:
-                run_buffer += "\n"
-
-        # Flush the final buffer and stdout after frame processing
-        sys.stdout.write(TerminalUtils.cursor_to_origin() + run_buffer)
-        sys.stdout.flush()
-        # Reset the run buffer
-        run_buffer = ""
+        # Output the accumulated buffer to stdout
+        self.write_to_term(buffer.getvalue())
+        self.flush_to_term()