… docs/readme_assets/optimization-before-after.gif
… docs/readme_assets/optimization_before_after.mp4
… examples/assets/nyx/alien-planet.nyx
… examples/assets/nyx/doom/doom-1-frame1.nyx
… examples/assets/nyx/doom/doom-1-frame10.nyx
… examples/assets/nyx/doom/doom-1-frame11.nyx
… examples/assets/nyx/doom/doom-1-frame12.nyx
… examples/assets/nyx/doom/doom-1-frame13.nyx
… examples/assets/nyx/doom/doom-1-frame14.nyx
… examples/assets/nyx/doom/doom-1-frame15.nyx
… examples/assets/nyx/doom/doom-1-frame16.nyx
… examples/assets/nyx/doom/doom-1-frame17.nyx
… examples/assets/nyx/doom/doom-1-frame18.nyx
… examples/assets/nyx/doom/doom-1-frame19.nyx
… examples/assets/nyx/doom/doom-1-frame2.nyx
… examples/assets/nyx/doom/doom-1-frame20.nyx
… examples/assets/nyx/doom/doom-1-frame21.nyx
… examples/assets/nyx/doom/doom-1-frame22.nyx
… examples/assets/nyx/doom/doom-1-frame23.nyx
… examples/assets/nyx/doom/doom-1-frame24.nyx
… examples/assets/nyx/doom/doom-1-frame25.nyx
… examples/assets/nyx/doom/doom-1-frame26.nyx
… examples/assets/nyx/doom/doom-1-frame27.nyx
… examples/assets/nyx/doom/doom-1-frame28.nyx
… examples/assets/nyx/doom/doom-1-frame29.nyx
… examples/assets/nyx/doom/doom-1-frame3.nyx
… examples/assets/nyx/doom/doom-1-frame30.nyx
… examples/assets/nyx/doom/doom-1-frame31.nyx
… examples/assets/nyx/doom/doom-1-frame32.nyx
… examples/assets/nyx/doom/doom-1-frame33.nyx