# [FEATURE]  Optimize `HemeraTermFx` terminal printing (major improvement)

## Overview:
`HemeraTermFx` had several high-penalty operations within its frame-to-string conversion method `generate_string_buffer`. Careful refactoring has improved frame printing time, at maximum supported resolution of 640*480, from 0.95ms to 0.0069ms. This is the equivalent of ~8.5fps -> ~45fps.

### Checklist:
- [x] My code follows the project's style guidelines.
- [x] I have made corresponding changes to the documentation.
- [ ] ~I have added tests that prove my changes are effective.~
- [ ] ~All new and existing tests pass.~

Note: Tests are still pending greater overall architecture maturity/stability.

## Summary of optimizations:
- Eliminate tuple creation for foreground and background color pairs by storing their sum in the delta frame matrix's background channel.
- Add row summation to skip processing entire rows with no changes.
- Skip processing individual pixels if their sum value is zero.
- Compute background color inline for changed pixels to avoid additional memory accesses.
- Pre-compute and reuse variables and comparison values outside loops.
- Replace string concatenations with list operations and io.StringIO for better performance.
- Remove row-level booleans by relying on implicit checks for row updates.
-Reduce redundant caching by avoiding unnecessary processing of "null" pixels.

## Other changes:
- Clean up and add stashed GIF demo files to `examples` folder.

Closes #26